more pkg cleanup
